+++ lib/pusher-fake/configuration.rb
@@ -47,8 +47,7 @@ module PusherFake
     def to_options(options = {})
       options.merge(
         wsHost: socket_options[:host],
-        wsPort: socket_options[:port],
-        enabledTransports: ["ws"]
+        wsPort: socket_options[:port]
       )
     end