+++ CHANGELOG.markdown
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+* Warn when Pusher configuration is not set. (Tristan Dunn)
+
 ## 1.4.0 — May 20th, 2015
 
 * Update development and test dependencies. (Tristan Dunn)