+++ CHANGELOG.markdown
@@ -4,6 +4,7 @@
 
 * Warn when Pusher configuration is not set. (Tristan Dunn)
 * Update the Pusher JS client to version 3.0.0. (Tristan Dunn)
+* Update development and test dependencies. (Tristan Dunn)
 
 ## 1.4.0 — May 20th, 2015